Googlehackathon BE/fastapi-service/requirements.txt
# Core ADK dependencies
google-adk
google-genai

# Web framework for streaming server
fastapi
uvicorn[standard]
websockets

# Environment and utilities
python-dotenv
certifi
